% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{gradient_palette_range}
\alias{gradient_palette_range}
\title{Función de gradiente con rango amplio}
\usage{
gradient_palette_range(n)
}
\value{
Función colorRampPalette
}
\description{
Función de gradiente con rango amplio
}
