% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{palette_ecpol}
\alias{palette_ecpol}
\title{Obtener una paleta de colores específica}
\usage{
palette_ecpol(name, n = NULL)
}
\arguments{
\item{name}{Nombre de la paleta a obtener}

\item{n}{Número de colores a devolver (opcional)}
}
\value{
Un vector de colores con clase "palette"
}
\description{
Obtener una paleta de colores específica
}
