% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{palette_four}
\alias{palette_four}
\title{Paleta de cuatro colores}
\format{
An object of class \code{palette} of length 4.
}
\usage{
palette_four
}
\value{
Vector de colores
}
\description{
Paleta de cuatro colores
}
\keyword{datasets}
