% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{palette_greenred}
\alias{palette_greenred}
\title{Paleta verde-rojo}
\format{
An object of class \code{palette} of length 2.
}
\usage{
palette_greenred
}
\value{
Vector de colores
}
\description{
Paleta verde-rojo
}
\keyword{datasets}
