% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{palette_one}
\alias{palette_one}
\title{Paleta de un solo color}
\format{
An object of class \code{palette} of length 1.
}
\usage{
palette_one
}
\value{
Vector de colores
}
\description{
Paleta de un solo color
}
\keyword{datasets}
