% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{palette_redgreen}
\alias{palette_redgreen}
\title{Paleta rojo-verde}
\format{
An object of class \code{palette} of length 2.
}
\usage{
palette_redgreen
}
\value{
Vector de colores
}
\description{
Paleta rojo-verde
}
\keyword{datasets}
