% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{palette_semaforo}
\alias{palette_semaforo}
\title{Paleta tipo semáforo}
\format{
An object of class \code{palette} of length 3.
}
\usage{
palette_semaforo
}
\value{
Vector de colores
}
\description{
Paleta tipo semáforo
}
\keyword{datasets}
