% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{palette_semaforoinverted}
\alias{palette_semaforoinverted}
\title{Paleta tipo semáforo invertida}
\format{
An object of class \code{palette} of length 3.
}
\usage{
palette_semaforoinverted
}
\value{
Vector de colores
}
\description{
Paleta tipo semáforo invertida
}
\keyword{datasets}
