% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{palette_three}
\alias{palette_three}
\title{Paleta de tres colores}
\format{
An object of class \code{palette} of length 3.
}
\usage{
palette_three
}
\value{
Vector de colores
}
\description{
Paleta de tres colores
}
\keyword{datasets}
