% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{palette_two}
\alias{palette_two}
\title{Paleta de dos colores}
\format{
An object of class \code{palette} of length 2.
}
\usage{
palette_two
}
\value{
Vector de colores
}
\description{
Paleta de dos colores
}
\keyword{datasets}
