% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{palettes_ecpol}
\alias{palettes_ecpol}
\title{Paletas de colores de Mais Enerxia}
\format{
Una lista con vectores de colores nombrados.
}
\usage{
palettes_ecpol
}
\description{
Definición de paletas de colores para uso en visualizaciones.
Incluye combinaciones discretas y de gradiente.
}
\keyword{datasets}
