% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\docType{data}
\name{print_gradient}
\alias{print_gradient}
\title{Matriz de ejemplo para mostrar gradientes}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 20 rows and 20 columns.
}
\usage{
print_gradient
}
\value{
Matriz numérica
}
\description{
Matriz de ejemplo para visualizar gradientes de color
}
\keyword{datasets}
