% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Palettes.R
\name{print_palette}
\alias{print_palette}
\title{Imprimir una paleta de colores}
\usage{
print_palette(x, ...)
}
\arguments{
\item{x}{Paleta de colores a imprimir}

\item{...}{Argumentos adicionales (no utilizados)}
}
\value{
No devuelve valor, imprime gráficamente la paleta
}
\description{
Imprimir una paleta de colores
}
