% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_plot.R
\name{save_to_pdf}
\alias{save_to_pdf}
\title{Funciones de exportación de graficos de EsadeEcPol}
\usage{
save_to_pdf(
  plot,
  filename = "",
  width = 18,
  height = 12,
  unit = "cm",
  dpi = 300
)
}
\arguments{
\item{plot}{componente al que asignamos la figura de ggplot}

\item{filename}{Nombre que se le quiere dar al archivo, incluyendo ruta}

\item{width}{Ancho de la figura}

\item{height}{Alto de la figura}

\item{unit}{Unidades de medida}

\item{dpi}{Densidad de pixeles por pulgada}
}
\value{
png and svg plots
}
\description{
Las funciones de este fichero sirven para hacer mas rapido la exportacion de las visualizaciones de Esade, ya que dejamos valores por defecot
}
